+++ lib/service/FirebaseMessaging.dart
@@ -50,6 +50,7 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging implements Notificatio
   static const String notifyPollOpen                                   = "$notifyBase.poll.create";
   static const String notifyEventsNotification                         = "$notifyBase.events";
   static const String notifyEventDetail                                = "$notifyBase.event.detail";
+  static const String notifyEventAttendeeSurveyInvitation              = "$notifyBase.event.attendee.survey.invitation";
   static const String notifyGameDetail                                 = "$notifyBase.game.detail";
   static const String notifyAthleticsGameStarted                       = "$notifyBase.athletics_game.started";
   static const String notifyAthleticsNewsUpdated                       = "$notifyBase.athletics.news.updated";
@@ -183,6 +184,7 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging implements Notificatio
   static const String payloadTypeOpenPoll = 'poll_open';
   static const String payloadTypeEvents = 'events';
   static const String payloadTypeEventDetail = 'event_detail';
+  static const String payloadTypeEvent = 'event';
   static const String payloadTypeGameDetail = 'game_detail';
   static const String payloadTypeAthleticsGameStarted = 'athletics_game_started';
   static const String payloadTypeAthleticsNewDetail = 'athletics_news_detail';
@@ -366,6 +368,13 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging implements Notificatio
     }
     else if (type == payloadTypeEventDetail) {
       NotificationService().notify(notifyEventDetail, data);
+    }
+    else if (type == payloadTypeEvent) {
+      String? entityType = JsonUtils.stringValue(data?['entity_type']);
+      String? operation = JsonUtils.stringValue(data?['operation']);
+      if ((entityType == 'event_attendance') && (operation == 'survey_invite')) {
+        NotificationService().notify(notifyEventAttendeeSurveyInvitation, data);
+      }
     }
     else if (type == payloadTypeGameDetail) {
       NotificationService().notify(notifyGameDetail, data);