+++ lib/ui/events2/Event2DetailPanel.dart
@@ -13,9 +13,11 @@ import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/events2/Event2AttendanceTakerPanel.dart';
 import 'package:illinois/ui/events2/Event2CreatePanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
+import 'package:illinois/ui/events2/Event2QrCodePanel.dart';
 import 'package:illinois/ui/events2/Event2SetupAttendancePanel.dart';
 import 'package:illinois/ui/events2/Event2SetupRegistrationPanel.dart';
 import 'package:illinois/ui/events2/Event2SetupSurveyPanel.dart';
+import 'package:illinois/ui/events2/Event2SurveyResponsesPanel.dart';
 import 'package:illinois/ui/events2/Event2Widgets.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
@@ -55,6 +57,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   Event2? _event;
   Survey? _survey;
+  Event2PersonsResult? _persons;
 
   // Keep a copy of the user position in the State because it gets cleared somehow in the widget
   // when sending the appliction to background in iOS.
@@ -249,7 +252,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
       ...?_adminSettingsButtonWidget,
       ...?_attendanceDetailWidget,
       ...?_contactsDetailWidget,
-      ...?_surveyDetailWidget,
+      ...?_detailsInfoWidget,
     ];
 
     return detailWidgets.isNotEmpty ? Padding(padding: EdgeInsets.only(top: 16), child:
@@ -380,20 +383,65 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     }
   }
 
-  List<Widget>? get _surveyDetailWidget {
-    if (_isParticipant && _hasSurvey) {
-      return <Widget>[
-        _buildTextDetailWidget(Localization().getStringEx('panel.event2.detail.survey.description', 'You will receive a notification with a follow up survey after this event.'), 'info',
-          textStyle: 'widget.info.regular.thin.italic', iconPadding: const EdgeInsets.only(right: 6),
-          maxLines: 3,
-        ),
-        _detailSpacerWidget
-      ];
-    } else {
-      return null;
+  List<Widget>? get _detailsInfoWidget {
+    
+    String? description;
+    bool hasRegistration = _event?.registrationDetails?.requiresRegistration ?? false;
+    bool hasAttendance = _event?.attendanceDetails?.isNotEmpty ?? false;
+    bool hasSurvey = (_event?.surveyDetails?.isNotEmpty ?? false) && (_survey != null);
+    bool showSurvey = (_isAttendee || _isAdmin) && hasAttendance && hasSurvey;
+    int surveyHours = _event?.surveyDetails?.hoursAfterEvent ?? 0;
+
+    if (hasRegistration) {
+      if (hasAttendance) {
+        if (showSurvey) {
+          if (_isAdmin) {
+            description = Localization().getStringEx('panel.event2.detail.survey.description.reg.att.svy.admin', 'This event requires Registering, Attendance will be taken and a Follow up Survey is set.');
+          }
+          else switch (surveyHours) {
+            case 0:  description = Localization().getStringEx('panel.event2.detail.survey.description.reg.att.svy.none', 'This event requires Registering, Attendance will be taken and you will receive a Notification with a Follow up Survey after this event.'); break;
+            case 1:  description = Localization().getStringEx('panel.event2.detail.survey.description.reg.att.svy.single', 'This event requires Registering, Attendance will be taken and you will receive a Notification with a Follow up Survey 1 hour after this event.'); break;
+            default: description = Localization().getStringEx('panel.event2.detail.survey.description.reg.att.svy.multi', 'This event requires Registering, Attendance will be taken and you will receive a Notification with a Follow up Survey {{hours}} hours after this event.').replaceAll('{{hours}}', surveyHours.toString()); break;
+          }
+        }
+        else {
+          description = Localization().getStringEx('panel.event2.detail.survey.description.reg.att', 'This event requires Registering and Attendance will be taken.'); 
+        }
+      }
+      else {
+        description = Localization().getStringEx('panel.event2.detail.survey.description.reg', 'Registration is required for this event.'); 
+      }
     }
+    else {
+      if (hasAttendance) {
+        if (showSurvey) {
+          if (_isAdmin) {
+            description = Localization().getStringEx('panel.event2.detail.survey.description.att.svy.admin', 'Attendance will be taken at this event and a Follow up Survey is set.');
+          }
+          else switch (surveyHours) {
+            case 0:  description = Localization().getStringEx('panel.event2.detail.survey.description.att.svy.none', 'Attendance will be taken at this event and you will receive a Notification with a Follow up Survey after this event.'); break;
+            case 1:  description = Localization().getStringEx('panel.event2.detail.survey.description.att.svy.single', 'Attendance will be taken at this event and you will receive a Notification with a Follow up Survey 1 hour after this event.'); break;
+            default: description = Localization().getStringEx('panel.event2.detail.survey.description.att.svy.multi', 'Attendance will be taken at this event and you will receive a Notification with a Follow up Survey {{hours}} hours after this event.').replaceAll('{{hours}}', surveyHours.toString()); break;
+          }
+        }
+        else {
+          description = Localization().getStringEx('panel.event2.detail.survey.description.att', 'Attendance will be taken at this event.'); 
+        }
+      }
+      else {
+        // No registration or attendance
+      }
+    }
+
+    return (description != null) ?<Widget>[
+      _buildTextDetailWidget(description, 'info',
+        textStyle: 'widget.info.regular.thin.italic',
+        iconPadding: const EdgeInsets.only(right: 6),
+        maxLines: 5,
+      ),
+      _detailSpacerWidget
+    ] : null;
   }
-  
 
   List<Widget>? get _contactsDetailWidget{
     if(CollectionUtils.isEmpty(_event?.contacts))
@@ -471,7 +519,12 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   }
 
   List<Widget>? get _registrationButtonWidget{
-    if(Auth2().isLoggedIn == false) //We can register only if logged in
+    if (Auth2().isLoggedIn == false) //We can register only if logged in
+      return null;
+
+    //Do not register/unregoster to past events
+    DateTime? startTimeUtc = _event?.startTimeUtc;
+    if ((startTimeUtc != null) && DateTime.now().toUtc().isAfter(startTimeUtc))
       return null;
 
     if (_event?.registrationDetails?.type == Event2RegistrationType.internal) { //Require App registration
@@ -480,7 +533,6 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
               title: Localization().getStringEx('panel.event2_detail.button.unregister.title', 'Unregister me'),
               onTap: _onUnregister,
               externalLink: false,
-              enabled: false,
               progress: _registrationLoading
           )];
         } else if (_event?.userRole == null){//Not registered yet
@@ -505,7 +557,10 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   }
 
   List<Widget>? get _followUpSurveyButtonWidget{
-    if (Auth2().isLoggedIn && (_event?.userRole == Event2UserRole.participant) && (_survey != null) && (_event?.isSurveyAvailable ?? false)) {
+    bool loggedIn = Auth2().isLoggedIn;
+    bool attendee = _isAttendee;
+    bool surveyAvailable = _event?.isSurveyAvailable ?? false;
+    if (loggedIn && attendee && (_survey != null) && (surveyAvailable)) {
       return <Widget>[_buildButtonWidget(
           title: Localization().getStringEx('panel.event2_detail.button.follow_up_survey.title', 'Take Survey'),
           onTap: _onFollowUpSurvey,
@@ -524,6 +579,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
           _buildSettingButton(title: "Event registration", onTap: _onSettingEventRegistration),
           _buildSettingButton(title: "Event attendance", onTap: _onSettingAttendance),
           _buildSettingButton(title: "Event follow-up survey", onTap: _onSettingSurvey),
+          _buildSettingButton(title: _survey != null ? "Event follow-up survey responses" : null, onTap: _onSettingSurveyResponses),
           _buildSettingButton(title: "Delete event", onTap: _onSettingDeleteEvent),
         ],)
     );
@@ -650,58 +706,26 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   void _onRegister() {
     Analytics().logSelect(target: 'Register me');
-
-    if ((_eventId != null) && !_registrationLoading) {
-        setStateIfMounted(() {
-          _registrationLoading = true;
-        });
-
-      Events2().registerToEvent(_eventId!).then((result) {
-        if (mounted) {
-            
-          if (result == true) {
-            Events2().loadEvent(_eventId!).then((Event2? event) {
-              if (mounted) {
-                setState(() {
-                  if (event != null) {
-                    _event = event;
-                  }
-                  _registrationLoading = false;
-                });
-              }
-            });
-          }
-          else {
-            setState(() {
-              _registrationLoading = false;
-            });
-            Event2Popup.showErrorResult(context, result);
-          }
-        }
-      });
-    }
+    _performRegistration(Events2().registerToEvent);
   }
 
   void _onUnregister() {
     Analytics().logSelect(target: 'Unregister me');
+    _performRegistration(Events2().unregisterFromEvent);
+  }
+
+  void _performRegistration(Future<dynamic> Function(String eventId) registrationApi) {
     if ((_eventId != null) && !_registrationLoading) {
         setStateIfMounted(() {
           _registrationLoading = true;
         });
 
-      Events2().unregisterFromEvent(_eventId!).then((result) {
+      registrationApi(_eventId!).then((result) {
         if (mounted) {
-            
-          if (result == true) {
-            Events2().loadEvent(_eventId!).then((Event2? event) {
-              if (mounted) {
-                setState(() {
-                  if (event != null) {
-                    _event = event;
-                  }
-                  _registrationLoading = false;
-                });
-              }
+          if (result is Event2) {
+            setState(() {
+              _event = result;
+              _registrationLoading = false;
             });
           }
           else {
@@ -778,16 +802,21 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   void _onSettingEditEvent(){
     Analytics().logSelect(target: "Edit event");
-    Navigator.push<Event2?>(context, CupertinoPageRoute(builder: (context) => Event2CreatePanel(event: _event, survey: _survey,))).then((Event2? result) {
-      if ((result != null) && mounted) {
-        setState(() {
-          _event = result;
+    Navigator.push<Event2SetupSurveyParam?>(context, CupertinoPageRoute(builder: (context) => Event2CreatePanel(event: _event, survey: _survey,))).then((Event2SetupSurveyParam? result) {
+      if (result != null) {
+        setStateIfMounted(() {
+          if (result.event != null) {
+            _event = result.event;
+          }
+          _survey = result.survey;
         });
       }
     });
   }
 
   void _onSettingPromote(){
+    Analytics().logSelect(target: "Promote Event", attributes: _event?.analyticsAttributes);
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => Event2QrCodePanel(event: _event)));
   }
 
   void _onSettingEventRegistration(){
@@ -833,6 +862,14 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     });
   }
 
+  void _onSettingSurveyResponses() {
+    Analytics().logSelect(target: "Event Survey Responses");
+    Navigator.push<Event2SetupSurveyParam?>(context, CupertinoPageRoute(builder: (context) => Event2SurveyResponsesPanel(
+      surveyId: _survey?.id,
+      eventName: _event?.name,
+    )));
+  }
+
   void _onSettingDeleteEvent(){
     Analytics().logSelect(target: 'Delete Event');
 
@@ -874,36 +911,63 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   Future<void> _initEvent() async {
     
-    Event2? event;
-    if ((_event == null) && StringUtils.isNotEmpty(widget.eventId) && mounted) {
+    String? eventId = widget.event?.id ?? widget.eventId;
+
+    Event2? event, theEvent;
+    if (_event != null) {
+      theEvent = _event;
+    }
+    else if (StringUtils.isNotEmpty(widget.eventId) && mounted) {
+      // Show loading progress only if we need to load the event.
       setState(() {
         _eventLoading = true;
       });
-      event = await Events2().loadEvent(widget.eventId!);
-    }
-
-    Survey? survey;
-    Event2? workEvent = event ?? _event;
-    if ((_survey == null) && (workEvent?.surveyDetails?.isNotEmpty ?? false) && StringUtils.isNotEmpty(workEvent?.id) && mounted) {
-      setState(() {
-        _eventLoading = true;
-      });
-      survey = await Surveys().loadEvent2Survey(workEvent!.id!);
-    }
-
-    setStateIfMounted(() {
-      _eventLoading = false;
-      if (event != null) {
+      theEvent = event = await Events2().loadEvent(widget.eventId!);
+      setStateIfMounted(() {
         _event = event;
+        _eventLoading = false;
+      });
+    }
+
+    // We need the survey and persons only if event has a servey attached.
+    if ((eventId != null) &&
+        (theEvent?.attendanceDetails?.isNotEmpty ?? false) &&
+        (theEvent?.surveyDetails?.isNotEmpty ?? false) &&
+        mounted) {
+      
+      List<Future<dynamic>> futures = [];
+
+      int? surveyIndex = (_survey == null) ? futures.length : null;
+      if (surveyIndex != null) {
+        futures.add(Surveys().loadEvent2Survey(eventId));
       }
-      if (survey != null) {
-        _survey = survey;
+
+      int? peopleIndex = (_persons == null) ? futures.length : null;
+      if (peopleIndex != null) {
+        futures.add(Events2().loadEventPeople(eventId));
       }
-    });
+
+      if (futures.isNotEmpty) {
+        List<dynamic> results = await Future.wait(futures);
+        Survey? survey = ((surveyIndex != null) && (surveyIndex < results.length) && (results[surveyIndex] is Survey)) ? results[surveyIndex] : null;
+        Event2PersonsResult? persons = ((peopleIndex != null) && (peopleIndex < results.length) && (results[peopleIndex] is Event2PersonsResult)) ? results[peopleIndex] : null;
+        setStateIfMounted(() {
+          _eventLoading = false;
+          if (persons != null) {
+            _persons = persons;
+          }
+          if (survey != null) {
+            _survey = survey;
+          }
+        });
+      }
+    }
   }
 
   Future<void> _refreshEvent({void Function(bool)? progress}) async {
-    if ((_eventId != null) && mounted) {
+    String? eventId = widget.event?.id ?? widget.eventId;
+    if ((eventId != null) && mounted) {
+      
       if (progress != null) {
         setState(() {
           progress(true);
@@ -911,18 +975,46 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
       }
 
       List<Future<dynamic>> futures = [
-        Events2().loadEvent(_eventId!)
+        Events2().loadEvent(eventId),
       ];
-      if (_event?.surveyDetails?.isNotEmpty ?? false) {
-        futures.add(Surveys().loadEvent2Survey(_eventId!));
+      
+      int? surveyIndex = ((_event?.attendanceDetails?.isNotEmpty ?? false) && (_event?.surveyDetails?.isNotEmpty ?? false)) ? futures.length : null;
+      if (surveyIndex != null) {
+        futures.add(Surveys().loadEvent2Survey(eventId));
       }
+
+      int? peopleIndex = ((_event?.attendanceDetails?.isNotEmpty ?? false) && (_event?.surveyDetails?.isNotEmpty ?? false)) ? futures.length : null;
+      if (peopleIndex != null) {
+        futures.add(Events2().loadEventPeople(eventId));
+      }
+
       List<dynamic> results = await Future.wait(futures);
       Event2? event = ((0 < results.length) && (results[0] is Event2)) ? results[0] : null;
-      Survey? survey = ((1 < results.length) && (results[1] is Survey)) ? results[1] : null;
+      Survey? survey = ((surveyIndex != null) && (surveyIndex < results.length) && (results[surveyIndex] is Survey)) ? results[surveyIndex] : null;
+      Event2PersonsResult? persons = ((peopleIndex != null) && (peopleIndex < results.length) && (results[peopleIndex] is Event2PersonsResult)) ? results[peopleIndex] : null;
+
+      futures.clear();
 
       // Handle the case if after refreshing event the surveyDetails get not empty
-      if ((event?.surveyDetails?.isNotEmpty ?? false) && (results.length < 2)) {
-        survey = await Surveys().loadEvent2Survey(_eventId!);
+      int? surveyIndex2 = ((surveyIndex == null) && (event?.attendanceDetails?.isNotEmpty ?? false) && (event?.surveyDetails?.isNotEmpty ?? false)) ? futures.length : null;
+      if (surveyIndex2 != null) {
+        futures.add(Surveys().loadEvent2Survey(eventId));
+      }
+
+      // Handle the case if after refreshing event the registrationDetails/attendanceDetails require loading persons
+      int? peopleIndex2 = ((peopleIndex == null) && (event?.attendanceDetails?.isNotEmpty ?? false) && (_event?.surveyDetails?.isNotEmpty ?? false)) ? futures.length : null;
+      if (peopleIndex2 != null) {
+        futures.add(Events2().loadEventPeople(eventId));
+      }
+
+      if (futures.isNotEmpty) {
+        results = await Future.wait(futures);
+        if ((peopleIndex2 != null) && (peopleIndex2 < results.length) && (results[peopleIndex2] is Event2PersonsResult)) {
+          persons = results[peopleIndex2];
+        }
+        if ((surveyIndex2 != null) && (surveyIndex2 < results.length) && (results[surveyIndex2] is Survey)) {
+          survey = results[surveyIndex2];
+        }
       }
 
       setStateIfMounted(() {
@@ -931,6 +1023,9 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
         }
         if (event != null) {
           _event = event;
+        }
+        if (persons != null) {
+          _persons = persons;
         }
         if (survey != null) {
           _survey = survey;
@@ -950,10 +1045,9 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   //Event getters
   bool get _isAdmin =>  _event?.userRole == Event2UserRole.admin;
   bool get _isAttendanceTaker =>  _event?.userRole == Event2UserRole.attendanceTaker;
-  bool get _isParticipant =>  _event?.userRole == Event2UserRole.participant;
+  //bool get _isParticipant =>  _event?.userRole == Event2UserRole.participant;
+  bool get _isAttendee => (_persons?.attendees?.indexWhere((person) => person.identifier?.accountId == Auth2().accountId) ?? -1) > -1;
 
-  bool get _hasSurvey => (_survey != null);
-  
   String? get _eventId => widget.event?.id ?? widget.eventId;
 
 }